Don't error on unsymbolizable query param keys
To fix https://errbit.publishing.service.gov.uk/apps/530f56010da115868600173d/problems/577c047a65786352db2b0400

This error occurs when canonicalizing query strings.  It appears that not every
character emerges unscathed from this process.  In this case, having `%E2`
(which is equivalent to "â")  in a key of a query string results in the error
found in Errbit.

The code path which breaks consists of the query string being parsed into a
hash:

`CGI::parse(self.query)`

(This converts `%E2` to `\xE2`)

`canonicalize_query!` then calls `to_sym` on this value, which throws
`EncodingError: invalid encoding symbol`

Our method for fixing this is to invert the comparison process, converting all
the allowed_query values to strings and comparing these with the string value
of each query hash key.  (Previously it would compare each value, then
symbolize it and attempt to compare it again.)

See https://trello.com/c/ihTQRRMX/325-encoding-error-on-hits-import-in-transition
for further explanation and context.