Fix "should not allow other schemas" test
This test was in a `lambda` and would pass regardless of whether
we set it `to raise_error` or `to_not raise_error`.
I've added a scheme validation method to the BLURI module, which
should only allow URIs beginning with http/https/mailto (the
latter is covered by tests elsewhere in this file).

We could split hairs on whether or not to return an `ArgumentError`
or a `URI::InvalidURIError`, but for the purposes of our usage,
anything with a different schema is not a valid URI. I see no
value in maintaining the distinction.

This commit should fix a resurgence in Sentry errors we're seeing
in Bouncer, which relies on optic14n to sanitize its URLs:
https://sentry.io/organizations/govuk/issues/2086648123/?project=202210